+++ matholymp/datasource.py
@@ -136,8 +136,11 @@ class DataSource(object):
         """
         Return the value of an attribute for a CountryEvent.  Possible
         names are: annual_url, code, name, flag_url, flag_filename,
-        is_official, is_normal, contact_emails, _person_ids,
-        _guide_ids, generic_id.
+        is_official, is_normal, contact_emails, expected_leaders,
+        expected_deputies, expected_contestants, expected_observers_a,
+        expected_observers_b, expected_observers_c,
+        expected_single_rooms, expected_numbers_confirmed,
+        _person_ids, _guide_ids, generic_id.
         """
         raise NotImplementedError